add notification badge and hover
