+++ components/modals/actorModal.tsx
@@ -1,0 +1,6 @@
+export const ActorModal = () => {
+  return (
+    <div className={"bg-white w-[100px] h-[107px] fixed right-0"}>
+    </div>
+  );
+};
